Migrate to jest, refactor file structure
