+++ .github/workflows/build-lint-test.yml
@@ -22,7 +22,7 @@ jobs:
       - run: yarn allow-scripts
       - run: yarn build
       - run: yarn lint
-      - run: yarn test:nobuild
+      - run: yarn test
   all-jobs-pass:
     name: All jobs pass
     runs-on: ubuntu-20.04
